requirements.txt
fastapi
uvicorn[standard]
mangum  # ASGI adapter for serverless (AWS Lambda, Vercel)
garminconnect
python-multipart

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
python-dateutil
psycopg2-binary>=2.9.0  # PostgreSQL driver for Vercel Postgres

# Hevy API integration
hevy-api-client
requests

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# CLI
click>=8.1.0
rich>=13.0.0

# Environment variables
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Date/time handling
pytz
